requirements.txt
kagglehub
python-dotenv
absl-py